feat: enhance add expense functionality and UI
- Added new context and provider for managing expense state.
- Introduced new components for handling expense addition, including form validation and dynamic item management.
- Updated routing and layout for better user experience.
- Integrated Playwright for end-to-end testing of the new features.
