+++ src/components/ui/checkbox.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 import * as CheckboxPrimitive from "@radix-ui/react-checkbox"
-import { CheckIcon } from "lucide-react"
+import CheckIcon from "lucide-react/dist/esm/icons/check"
 
 import { cn } from "@/lib/utils"
