+++ src/components/ui/select.tsx
@@ -1,6 +1,8 @@
 import * as React from "react"
 import * as SelectPrimitive from "@radix-ui/react-select"
-import { CheckIcon, ChevronDownIcon, ChevronUpIcon } from "lucide-react"
+import CheckIcon from "lucide-react/dist/esm/icons/check"
+import ChevronDownIcon from "lucide-react/dist/esm/icons/chevron-down"
+import ChevronUpIcon from "lucide-react/dist/esm/icons/chevron-up"
 
 import { cn } from "@/lib/utils"
