+++ src/routes/event-result/types.ts
@@ -1,0 +1,6 @@
+import type { EventType } from '@/types';
+
+export type EventResultPageProps = {
+  eventList: EventType[];
+  handleUpdateEventById: (data: EventType) => void;
+};
