+++ tests/screenshot-flow.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect } from '@playwright/test';
+import { test } from '@playwright/test';
 import { mkdirSync } from 'fs';
 import { join } from 'path';
